Fix: 메인 페이지 footer nav 파일 이름 변경 및 코드 이동
